restructured the project
