+++ threemf2-thumbnail/src/mesh_pipeline.rs
