--- threemf2-thumbnail/src/thumbnail/mod.rs
@@ -1,4 +1,0 @@
-mod thumbnail_generator;
-pub use thumbnail_generator::{ThumbnailConfig, ThumbnailGenerator};
-
-mod mesh_pipeline;
